task3 2nd commit

+++ server/src/main/java/com/qualco/casestudy/dto/CountryDataDto.java
@@ -6,6 +6,6 @@ public record CountryDataDto(
     String region,
     String country,
     Integer year,
-    Long population,
-    Long gdp
+    Integer population,
+    Double gdp
 ) {}
